fix: modal footer and image slidder
